# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.ErrorDetail:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  dto.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/dto.ErrorDetail'
      success:
        type: boolean
    type: object
...
  dto.QueueMinimalInfo:
    properties:
      id:
        type: string
      status:
        type: string
      subject:
        $ref: '#/definitions/dto.SubjectResponse'
    type: object
...
  dto.SlotResponse:
    properties:
      id:
        type: string
      position:
        type: integer
      queue:
        $ref: '#/definitions/dto.QueueMinimalInfo'
      queue_id:
        type: string
      signed_up_at:
        type: string
      status:
        type: string
      student:
        $ref: '#/definitions/dto.UserMinimalResponse'
    type: object
  dto.SubjectResponse:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  dto.SuccessResponse:
    properties:
      data: {}
      success:
        type: boolean
    type: object
...
  dto.TransferSlotsRequest:
    properties:
      source_queue_id:
        type: string
    required:
    - source_queue_id
    type: object
  dto.TransferSlotsResponse:
    properties:
      slots:
        items:
          $ref: '#/definitions/dto.SlotResponse'
        type: array
      transferred_count:
        type: integer
    type: object
...
  dto.UpdateSlotRequest:
    properties:
      status:
        type: string
    required:
    - status
    type: object
...
  dto.UserMinimalResponse:
    properties:
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
    type: object
...
info:
  contact: {}
  description: HTTP API for Stud Hub backend.
  title: Stud Hub API
  version: "1.0"
paths:
...
  /queues/{queue_id}/slots:
    get:
      description: Returns all slots in a queue.
      parameters:
      - description: Queue ID
        in: path
        name: queue_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.SlotResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: List queue slots
      tags:
      - Queue Slots
    post:
      description: Signs up authenticated user for queue and returns created slot
        data.
      parameters:
      - description: Queue ID
        in: path
        name: queue_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.SlotResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: Sign up for queue
      tags:
      - Queue Slots
  /queues/{queue_id}/slots/{slot_id}:
    delete:
      description: Cancels authenticated user's queue sign-up.
      parameters:
      - description: Queue ID
        in: path
        name: queue_id
        required: true
        type: string
      - description: Slot ID
        in: path
        name: slot_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: Cancel queue slot
      tags:
      - Queue Slots
    get:
      description: Returns slot details by slot ID.
      parameters:
      - description: Queue ID
        in: path
        name: queue_id
        required: true
        type: string
      - description: Slot ID
        in: path
        name: slot_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.SlotResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: Get queue slot by ID
      tags:
      - Queue Slots
    patch:
      consumes:
      - application/json
      description: Updates queue slot status.
      parameters:
      - description: Queue ID
        in: path
        name: queue_id
        required: true
        type: string
      - description: Slot ID
        in: path
        name: slot_id
        required: true
        type: string
      - description: Update slot request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateSlotRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.SlotResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: Update queue slot
      tags:
      - Queue Slots
  /queues/{queue_id}/transfers:
    post:
      consumes:
      - application/json
      description: Transfers failed students from source queue to target queue.
      parameters:
      - description: Target queue ID
        in: path
        name: queue_id
        required: true
        type: string
      - description: Transfer slots request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.TransferSlotsRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.TransferSlotsResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: Transfer failed slots
      tags:
      - Queue Slots
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
  CookieAuth:
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"